Massive CRUD updates for proficiency, frequency, critical, and classification

+++ database/migrations/2017_01_27_010452_create_frequency_table.php
@@ -14,7 +14,7 @@ class CreateFrequencyTable extends Migration
     public function up()
     {
         Schema::create('frequency', function (Blueprint $table) {
-            $table->smallInteger('id');
+            $table->increments('id');
             $table->string('name');
             $table->timestamps();
         });
